// src/model/vscodeModels.ts
// Copyright © 2025 Cisco Systems, Inc. and its affiliates.
// All rights reserved.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.


export interface VscodeCopilotChatMode {
  description: string;
  tools: string[];
  model: string;
  prompt: string;
}

export interface MCPServerConfiguration {
  name: string;
  type: string;
  command?: string;
  args?: string[];
  env?: Record<string, string>;
  url?: string;
  headers?: Record<string, string>;
  //  inputs?: InputConfig[];
}

interface InputConfig {
  id: string;
  type: "promptString" | "pickString" | "command";
  description?: string;
  default?: string;
  password?: boolean;
  options?: string[];
  args?: string[];
}